Fix and add Ngff test
