+++ src/test/java/projects/ngff/CYXNgffDataTest.java
@@ -36,8 +36,8 @@ public class CYXNgffDataTest extends BaseTest {
             throw new RuntimeException("Coordinates out of bounds");
         }
 
-        RandomAccessibleInterval<?> realPixelValue = spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(0).getImage(0);
-        VolatileCachedCellImg volatileCachedCellImg = (VolatileCachedCellImg) realPixelValue;
+        RandomAccessibleInterval<?> randomAccessibleInterval = spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(0).getImage(0);
+        VolatileCachedCellImg volatileCachedCellImg = (VolatileCachedCellImg) randomAccessibleInterval;
         CellGrid cellGrid = volatileCachedCellImg.getCellGrid();
         long[] dims = new long[]{1024, 930, 1};
         int[] cellDims = new int[]{256, 256, 1};
