+++ src/test/java/projects/ngff/YXNgffDataTest.java
@@ -10,6 +10,7 @@ import mpicbg.spim.data.SpimDataException;
 import net.imglib2.FinalDimensions;
 import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.img.cell.CellGrid;
+import net.imglib2.type.numeric.integer.UnsignedShortType;
 import projects.remote.BaseTest;
 
 @Slf4j
@@ -27,20 +28,39 @@ public class YXNgffDataTest extends BaseTest {
 
     @Test
     public void checkDataset() {
-        long x = 1;
-        long y = 1;
+        long x = 0;
+        long y = 0;
         long z = 1;
         long[] imageDimensions = spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(0).getImage(0).dimensionsAsLongArray();
         if (x > imageDimensions[0] || y > imageDimensions[1] || z > imageDimensions[2]) {
             throw new RuntimeException("Coordinates out of bounds");
         }
 
-        RandomAccessibleInterval<?> realPixelValue = spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(0).getImage(0);
-        VolatileCachedCellImg volatileCachedCellImg = (VolatileCachedCellImg) realPixelValue;
+        RandomAccessibleInterval<?> randomAccessibleInterval = spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(0).getImage(0);
+        VolatileCachedCellImg volatileCachedCellImg = (VolatileCachedCellImg) randomAccessibleInterval;
         CellGrid cellGrid = volatileCachedCellImg.getCellGrid();
         long[] dims = new long[]{1024, 930, 1};
         int[] cellDims = new int[]{256, 256, 1};
         CellGrid expected = new CellGrid(dims, cellDims);
         Assertions.assertEquals(expected, cellGrid);
+
+        UnsignedShortType o = (UnsignedShortType)  randomAccessibleInterval.getAt(0, 0, 0);
+        int im = o.get();
+        Assertions.assertEquals(538, im);
+    }
+
+    @Test
+    public void checkImgValue() {
+        long x = 0;
+        long y = 0;
+        long z = 1;
+        long[] imageDimensions = spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(0).getImage(0).dimensionsAsLongArray();
+        if (x > imageDimensions[0] || y > imageDimensions[1] || z > imageDimensions[2]) {
+            throw new RuntimeException("Coordinates out of bounds");
+        }
+        RandomAccessibleInterval<?> randomAccessibleInterval = spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(0).getImage(0);
+        UnsignedShortType o = (UnsignedShortType)  randomAccessibleInterval.getAt(0, 0, 0);
+        int value = o.get();
+        Assertions.assertEquals(538, value);
     }
 }
