+++ src/test/java/spimdata/OmeZarrV4S3SpimDataTests.java
@@ -1,23 +1,16 @@
 package spimdata;
 
 import java.io.IOException;
-import java.util.Arrays;
 
 import org.embl.mobie.io.ome.zarr.openers.OMEZarrS3Opener;
 import org.junit.jupiter.api.Test;
 
 import mpicbg.spim.data.SpimData;
-import mpicbg.spim.data.sequence.MultiResolutionSetupImgLoader;
-import net.imglib2.Cursor;
-import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.type.numeric.NumericType;
-import net.imglib2.type.numeric.RealType;
-import net.imglib2.view.Views;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class OmeZarrV4S3SpimDataTests<N extends NumericType<N> & RealType<N>> {
+public class OmeZarrV4S3SpimDataTests {
     public static final String ZYX_FILE_KEY = "https://s3.embl.de/i2k-2020/ngff-example-data/v0.4/zyx.ome.zarr";
     public static final String CZYX_FILE_KEY = "https://s3.embl.de/i2k-2020/ngff-example-data/v0.4/czyx.ome.zarr";
     public static final String CYX_FILE_KEY = "https://s3.embl.de/i2k-2020/ngff-example-data/v0.4/cyx.ome.zarr";
@@ -47,7 +40,8 @@ public class OmeZarrV4S3SpimDataTests<N extends NumericType<N> & RealType<N>> {
 
         for (int setupId = 0; setupId < numSetups; setupId++) {
             System.out.println("setup: " + setupId);
-            final Info info = getImgInfo(spimData, setupId);
+            Info info = new Info();
+            info = info.getImgInfo(spimData, setupId);
             info.print();
             assertArrayEquals(new long[]{128, 66, 122}, info.dimensions);
             assertEquals(3, info.levels);
@@ -58,46 +52,7 @@ public class OmeZarrV4S3SpimDataTests<N extends NumericType<N> & RealType<N>> {
         }
     }
 
-    private Info getImgInfo(SpimData spimData, int setupId) {
-        // TODO we could add a method getNumTimepoints() to our ImageLoader?
-        //   Then we could use this in the tests
-        final MultiResolutionSetupImgLoader<N> setupImgLoader = (MultiResolutionSetupImgLoader) spimData.getSequenceDescription().getImgLoader().getSetupImgLoader(setupId);
-        final int numMipMapLevels = setupImgLoader.numMipmapLevels();
-        final int level = numMipMapLevels - 1;
 
-        final RandomAccessibleInterval<N> image = setupImgLoader.getImage(0, level);
-        final Cursor<N> cursor = Views.iterable(image).cursor();
-        final Info info = new Info();
-        while (cursor.hasNext()) {
-            final N next = cursor.next();
-            if (next.getRealDouble() > info.max)
-                info.max = next.getRealDouble();
-            ;
-            if (next.getRealDouble() < info.min)
-                info.min = next.getRealDouble();
-            ;
-        }
-        info.dimensions = image.dimensionsAsLongArray();
-        info.level = level;
-        info.levels = numMipMapLevels;
 
-        return info;
-    }
 
-    class Info // for one setup
-    {
-        public int levels;
-        public int level; // lowest resolution level
-        double min = Double.MAX_VALUE; // at lowest resolution level
-        double max = -Double.MAX_VALUE; // at lowest resolution level
-        long[] dimensions; // at lowest resolution level
-
-        public void print() {
-            System.out.println("Levels: " + levels);
-            System.out.println("Lowest level: " + level);
-            System.out.println("Min: " + min);
-            System.out.println("Max: " + max);
-            System.out.println("Dimensions: " + Arrays.toString(dimensions));
-        }
-    }
 }
